Adding SyncDiff::DB - Database object
This object encapsulates the database operations for SyncDiff, as
well as acts as the control process for reading and writing to the
database.  Because most of SyncDiff is going to be occuring in chroot
environments this needs to fork off and be self maintaining.

Basically it forks off and creates a two way pipe between the main
syncdiff program and itself.  Interaction with the database is done
by calling public functions in the DB object.  The object then
internally creates a JSON object with the request and shoves it
across the IPC, decodes it and acts on it, eventually passing back
a response.  the JSON and IPC calls are basically hidden by the API
which I'd argue is a good thing.