Adding SyncDiff::Scanner - Scanner object
This is basically a wrapper to deal with chrooting and checking
existing files for file changes and dealing with updating the
database.  This forks and chroots itself, so it doesn't actually
have direct access to the database.

While it doesn't have it yet, it is intended to have the iNotify
code in here as well, but right now implements enough of file
scanning for `syncdiff --scan` to work.