DB: gethostbyname, or the annoyances of chrooting for security
So by chrooting I pull out the ability for the system to reference
any of it's DNS mechanisms, so it can't resolve hostnames to ip
addresses.  To get around this I just accept that hostname are now
looked up via the database (not chrooted) and pass it back.

It's kinda like a database so it sorta vaguely fits ;-)

It's a little ridiculous to do it this way but it does keep the
chrooted processes from needing extra things shoved into their filespace