syncdiff: Add in basic sync socket authentication
This should get us basic authentication on initial creation of the
connection from the client to the server.  This obviously doesn't
handle any ongoing authentication (I.E. an auth token of sorts) or
anything else like that, and there's obvious gotchas in just comparing
a shared secret.  For now this will work, but will likely need to be
revisited later.

This also means that config and various other pieces, need to get passed
around into the client and server modules, which this incorporates