SenderReceiver: convert to using plain_send() and plain_reciever()
The send and recieve code was starting to get all over the place
and copy/pasted around a lot.  Obviously a mess and so this is the
first round of cleaning that up.  There's some other minor modifications
that will need to get done to this from the server perspective but
this should clean it up for the client side of things.