SenderReceiver: Fixing some of my divergent thinking
Somewhere along the way SenderReceiver took a small fork.  The server
side apparently does more in-house processing of the returned stream
in the Protocol object.  The client however assumes it all happens
in the SenderReciever functions.

*DOH*

This resolves it for now, it's a bit of a hack, and I'll be the
first to admit it's stupid.  But this gets it out the door for now.